ARNAQ

ARNAQ stands for Analysts' RNA QC. ARNAQ is a tool for exploring RNA sequencing datasets, with an eye to differential expression or other analysis. It is designed to make reviewing the data and isolating potential QC issues as simple and swift as possible.

This function creates a report based on the RNA count data, genome resources and metadata resources, in html format. It uses external files to define specific file locations and parameters rather than incorporating them all into the function arguments.

arnaq(
  resources.file = "resources.yml",
  sample.file = "samples.txt",
  model.name = NULL,
  sample.mask = NULL,
  treat.groups = NULL,
  pca.groups = NULL,
  ERCC = FALSE,
  normalise = "VSN",
  svg.export = FALSE,
  max.samples.per.page = 40,
  pca.depth = NULL,
  ERCC.pairs = list(),
  ERCC.combined = TRUE,
  scatter.pairs = list(),
  gene.mask.name = "Genes"
)

Arguments

resources.file

The file containing project-centric data, including the location of counts tables and metric tables. The default looks for the file resources.yml in your working directory, and is useful when you have one project per directory.

sample.file

The file containing sample-centric data. The default looks for the file samples.txt in your working directory, and is useful when you have one project per directory.

model.name

An optional model name that allows you to have multiple ARNAQ reports in parallel for the same project. The model name is appended to the project name in ARNAQ outputs. Note that running ARNAQ will replace its generated data objects in the current R session.

sample.mask

A vector of True/False values, with length equal to the number of samples, indicating which samples to include in the report. This is used to run reports excluding outliers. \link{make_outlier_mask} is a convenient function for generating one. Default behaviour is to include all samples.

treat.groups

A vector of character objects that specifices which of the group columns in the samples.txt file are to be considered treatment groups for DESeq2's normalisation. If the group information is not of full rank, this should be a constrained set of these groups that is of full rank. Default behaviour is to use all groups.

pca.groups

A vector of character objects that specify which groups in samples.txt to create PCA plots for. Default behaviour is to only use those groups in treat.groups.

ERCC

A logical value indicating if ERCC spike-ins are included in the gene set.

normalise

A character value specifying which normalisation to perform on the data. VSN is the default, but this does not run in reaosnable time for large (>=30) numbers of samples. linear is the alternative. Any other value will skip normalisation.

svg.export

If set to TRUE, create an additional directory in the output called svg with svg versions of every plot created in the report, suitable for inclusion in papers.

max.samples.per.page

In projects with a large number of samples, some plots will be spread across multiple 'pages' in the output report, to keep them readable. Alter the default value to fine-tune the point at which this behaviour will occur.

pca.depth

How many components to use when preparing PCA plots. Components beyond 2 will be plotted against the first component, rather than showing every pairwise combination up to the limit. Default behaviour is to use all components that account for at least 10% of the variation, with a minimum of 2.

ERCC.pairs

A list of pairs of samples to produce ERCC-specific fold-change plots for. Each entry in the list should be a length 2 vector of character objects, naming those samples. Should be an empty list if ERCC spike-ins are not present in the data.

ERCC.combined

When TRUE, the ERCC.pairs plots above will use concentration as the x-axis, rather than the expected fold-change.

scatter.pairs

A list of pairs of samples to produce read counts scatter plots for, to check consistency between samples that are technical replicates.

gene.mask.name

Which set of genes to filter to in preparing the report. This should be a name that will be present in the gene.masks list. See the objects created vignette for a list of options. The default of Genes will use all genes in the reference .gtf with a non-zero count for at least one sample, that are not ERCC spike-ins.

Details

It is not a replacement for the first line QC performed by sequencing platforms and their associated tools; but instead to supplement them by considering experimental groups and allowing a more nuanced approach to potentially problematic samples than a simple pass/fail flag.

Analysts who handle a lot of RNA projects will benefit from being able to perform these QC steps more quickly, and will help avoid copy/paste scripts that are prone to errors where one line is not updated for the current project. The report contains descriptions of what the plots show, helping to present this work to collaborators who are less familiar with bioinformatics.

Investigators who have some familiarity with bioinformatics will benefit from shorter, clearer R scripts where they do not have to delve too deeply into the specifics of the packages used for this QC analysis, helping them be confident that the analysis they perform is reliable.

In both cases, the .svg export options the produces high quality print-ready plots that can be easily incorporated into papers.

ARNAQ is named in reference to the board game 'Lost Ruins of Arnak'.

resources.yml contains information that is identical for all samples, or refers to the project as a whole.

samples.txt contains sample-centric information, such as membership of an experimental group.

Both of these are assumed to remain unchanged through different ARNAQ runs. Arguments to this function are those options that may change run to run. More details are in ????.

A directory will be created in the current working directory called QC to house the output files, over-writing previous versions, if any.

As a side effect of running this function, a number of R objects are created in the session. The intention is that these are available for downstream analysis, but can also be adjusted (such as fine-tuning plots). The R console output will also appear in the file arnaq.log in your working directory.

ARNAQ will try to optimise by looking in the session for some data before attempting to read it from the file locations you give it. If the underlying files have changed, use \link{arnaq_clear} to remove ARNAQ's objects before rerunning.

See also

Author

Maintainer: Ben J Wright chthonic.games@gmail.com (ORCID)

Other contributors:

Helen Lockstone [contributor]

Eshita Sharma [contributor]

Santiago Revale [contributor]

Matthieu Miossec [contributor]

Duncan Parkes [contributor]

Examples

if (FALSE) {
arnaq()
arnaq("some/other/location/resources.yml", "some/other/location/samples.txt")
arnaq(model.name = "outliers removed", sample.mask = some_mask, normalise = "linear")
}