Exports all the plots created by arnaq in svg format.

This function is called automatically when \link{arnaq} is called with svg=TRUE. However, it can be called after adjusting any of the generated plots to ensure the svg export is of the correct versions of them.

export_all_svg(svg.directory = "QC/svg", qc.name = qc.name)

Arguments

svg.directory

The directory into which to make the export. If it does not exist, it will be created, if possible. The default is QC/svg.

qc.name

The prefix to use for all the file names. By default, this is the project name - underscore - model name as set in resources.yml and the model.name argument to \link{arnaq}.

See also