Create a plot showing complexity

The complexity is measured by the proportion of the reads in the top n genes, as n varies, per sample. Thinning is performed to reduce the file size for svg and html output).

plot_complexity_curves(
  complexity.mat,
  highlight = TRUE,
  threshold = 8,
  pivot = 10
)

Arguments

complexity.mat

The matrix of calculated complexity, created automatically as part of the main ARNAQ workflow.

highlight

When TRUE, highlighting will be performed. When FALSE, all samples will be drawn in colour, and will appear in the legend.

threshold

The number of 'worst' samples to highlight, out of the total.

pivot

What 'top n' to use when determining the worst samples.

Value

A plot showing proportion of reads assigned to the top n genes, for each sample, with the least complex samples highlighted.

Details

To improve readability, this function can select the least complex samples to highlight, and only show those in the legend. Non-highlighted samples will still have their curves drawn, just in grey behind the highlighted samples.

In order to determine the 'worst' samples, a representative top n is used.